requirement.txt
aiomysql==0.2.0
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.16.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
packaging==24.1
platformdirs==4.3.6
ply==3.11
PyJWT==2.9.0
PyMySQL==1.1.1
Pyomo==6.7.1
setuptools==69.5.1
virtualenv==20.26.5
Werkzeug==3.0.4
wheel==0.43.0
